Move AutoMigrationSpecFrom6to7.kt to new package
